新增 Metrics 概念，新增 PluginListener 等容器监听器修改，其他重构
1.新增 Metrics 概念
==>实现插件运行日志
    ==>MetrizablePlugin 实现插件指标信息的汇报
    ==>MetricPlugin 实现插件指标信息的收集和打印
    ==>Metrics 存储一个插件实例的指标信息
==>可通过 Spring Boot application.yml 响应属性配置
    ==>可配置打印间隔帧数
    ==>可配置打印插件类型
==>BaseMetricPlugin 实现默认 Metrics 收集和打印
    ==>复数个同 category 插件时自动编号
    ==>打印日志时插件 name 对齐
==>原有插件相应调整，实现 MetrizablePlugin 接口
    ==>MovablePlugin 等插件实现 delay 指标的汇报
    ==>UIPlugin 实现 refreshUI 指标的汇报

2.ContainerListener 修改
==>新增 PluginListener
    ==>用于实现对容器中 Plugin 的监听
    ==>PluginableContainer 中相应增加 add、remove 和 notify 监听器的逻辑
    ==>ListensPlugin 相应增加 PluginListener 类型监听器的 add 和 remove
==>ObjectListener 和 LifecycleListener 所有接口回调方法增加 Container 参数
    ==>表明发起回调的容器对象
    ==>用于支持一个监听器同时监听多个容器

3.其他重构
==>依赖 beginning-thread 重构倒计时实现方式

+++ ib-core/src/main/java/cn/milai/ib/util/ImageUtil.java
@@ -150,9 +150,9 @@ public abstract class ImageUtil {
 	 * @param transaparency
 	 * @return
 	 */
-	public static Graphics createGraphics(BufferedImage img, float transaparency) {
+	public static Graphics createGraphics(BufferedImage img, double transaparency) {
 		Graphics2D g2d = img.createGraphics();
-		g2d.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, transaparency));
+		g2d.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, (float) transaparency));
 		return g2d;
 	}
